This version supports domain controllers
The other version supported domain controllers too, but it had to be ran as the domain admin because of it's use of wmic. This version can be run as a "backup operators" user. This is not a replacement for the other version though, because this version will only work on server OSes since 'vssadmin Create Shadow' is not available to client OSes